Added generators, reworked menu a bit.

+++ com.evolveum.midpoint.eclipse.parent/com.evolveum.midpoint.eclipse.ui/src/com/evolveum/midpoint/eclipse/ui/handlers/server/ServerResponseItem.java
@@ -125,7 +125,7 @@ public abstract class ServerResponseItem<SR extends ServerResponse> {
 
 	private void removeFiles() {
 		for (IFile file : getFiles()) {
-			if (file.exists()) {
+			if (file != null && file.exists()) {
 				try {
 					file.delete(true, null);
 				} catch (CoreException e) {
